remove accidental duplicate EditorPlugin class.
