--- addons/RichLogger/Plugin.cs.uid
@@ -1,1 +1,0 @@
-uid://dyoe1vix8y1cn
