+++ addons/RichLogger/plugin.cfg
@@ -4,4 +4,4 @@ name="RichLogger"
 description="Logger with colored logs and stack traces. Configurable from Editor Output window."
 author="Piotr Przestrzelski dimzi"
 version="1.0.0"
-script="Plugin.cs"
+script="LoggerToolbarPlugin.cs"
